requirements.txt
pyspark
river
kafka-python